refactor: presentation層の残存ファイルをscenes/shared配下に移行（TASK-0105）
- KeyboardManagerをshared/services/inputに移動
- title/rankup/shopコンポーネントをscenes/componentsに移動
- 旧ファイルは後方互換性のため再エクスポートに変換
