+++ atelier-guild-rank/src/presentation/scenes/components/rankup/index.ts
@@ -1,12 +1,16 @@
 /**
- * RankUp コンポーネント バレルエクスポート
- * TASK-0055 RankUpSceneリファクタリング
+ * RankUp Components（後方互換性用再エクスポート）
+ *
+ * @description
+ * 実体は scenes/components/rankup/ に移動済み。
+ * 後方互換性のため再エクスポートを提供する。
+ *
+ * 新規コードでは @scenes/components/rankup を使用すること。
  */
-
-export { RankUpHeader } from './RankUpHeader';
-export { RankUpRequirements } from './RankUpRequirements';
-export { RankUpRewards } from './RankUpRewards';
-export { RankUpTestPanel } from './RankUpTestPanel';
+export { RankUpHeader } from '@scenes/components/rankup/RankUpHeader';
+export { RankUpRequirements } from '@scenes/components/rankup/RankUpRequirements';
+export { RankUpRewards } from '@scenes/components/rankup/RankUpRewards';
+export { RankUpTestPanel } from '@scenes/components/rankup/RankUpTestPanel';
 export type {
   Artifact,
   Quality,
@@ -16,4 +20,4 @@ export type {
   RankUpReward,
   TestPanelCallbacks,
   TestState,
-} from './types';
+} from '@scenes/components/rankup/types';
