+++ atelier-guild-rank/src/presentation/scenes/components/shop/ShopHeader.ts
@@ -1,148 +1,10 @@
 /**
- * ShopHeaderコンポーネント
- * TASK-0056 ShopSceneリファクタリング
+ * ShopHeader（後方互換性用再エクスポート）
  *
  * @description
- * ショップのヘッダー部分を表示するコンポーネント
- * - タイトル「ショップ」
- * - 所持金表示
- * - 戻るボタン
+ * 実体は scenes/components/shop/ShopHeader.ts に移動済み。
+ * 後方互換性のため再エクスポートを提供する。
+ *
+ * 新規コードでは @scenes/components/shop を使用すること。
  */
-
-import type { RexLabel } from '@presentation/types/rexui';
-import { BaseComponent } from '@presentation/ui/components/BaseComponent';
-import { Colors, THEME } from '@presentation/ui/theme';
-import type Phaser from 'phaser';
-import type { OnBackClickCallback } from './types';
-
-/** ShopHeaderオプション */
-interface ShopHeaderOptions {
-  /** 戻るボタンクリック時のコールバック */
-  onBackClick?: OnBackClickCallback;
-}
-
-/**
- * ShopHeaderコンポーネント
- * ショップ画面のヘッダー部分を担当
- */
-export class ShopHeader extends BaseComponent {
-  /** 所持金テキスト */
-  private goldText: Phaser.GameObjects.Text | null = null;
-
-  /** 現在の所持金 */
-  private currentGold: number = 0;
-
-  /** オプション */
-  private options: ShopHeaderOptions;
-
-  /**
-   * 戻るボタン
-   * TASK-0059: rexUI型定義を適用
-   */
-  private backButton: RexLabel | null = null;
-
-  /**
-   * コンストラクタ
-   * @param scene Phaserシーン
-   * @param x X座標
-   * @param y Y座標
-   * @param options オプション設定
-   */
-  constructor(scene: Phaser.Scene, x: number, y: number, options?: ShopHeaderOptions) {
-    super(scene, x, y);
-    this.options = options || {};
-  }
-
-  /**
-   * コンポーネントを作成
-   */
-  create(): void {
-    // タイトルテキスト
-    const titleText = this.scene.add.text(50, 20, 'ショップ', {
-      fontSize: `${THEME.sizes.xlarge}px`,
-      color: `#${Colors.text.primary.toString(16).padStart(6, '0')}`,
-    });
-    titleText.setOrigin(0, 0.5);
-    this.container.add(titleText);
-
-    // 所持金テキスト
-    this.goldText = this.scene.add.text(300, 20, `所持金: ${this.currentGold}G`, {
-      fontSize: `${THEME.sizes.medium}px`,
-      color: `#${Colors.text.accent.toString(16).padStart(6, '0')}`,
-    });
-    this.goldText.setOrigin(0, 0.5);
-    this.container.add(this.goldText);
-
-    // 戻るボタン
-    this.createBackButton();
-  }
-
-  /**
-   * 戻るボタンを作成
-   */
-  private createBackButton(): void {
-    if (!this.rexUI) {
-      return;
-    }
-
-    // rexUIラベルで戻るボタンを作成
-    this.backButton = this.rexUI.add.label({
-      x: 700,
-      y: 20,
-      width: 80,
-      height: 32,
-      background: this.scene.add
-        .graphics()
-        .fillStyle(Colors.ui.button.normal, 1)
-        .fillRoundedRect(0, 0, 80, 32, 4),
-      text: this.scene.add.text(0, 0, '戻る', {
-        fontSize: `${THEME.sizes.medium}px`,
-        color: `#${Colors.text.primary.toString(16).padStart(6, '0')}`,
-      }),
-      align: 'center',
-    });
-
-    this.backButton.setInteractive();
-    this.backButton.on('pointerdown', () => {
-      if (this.options.onBackClick) {
-        this.options.onBackClick();
-      }
-    });
-
-    this.container.add(this.backButton);
-  }
-
-  /**
-   * 所持金を設定
-   * @param gold 所持金
-   */
-  setGold(gold: number): void {
-    this.currentGold = gold;
-    if (this.goldText) {
-      this.goldText.setText(`所持金: ${gold}G`);
-    }
-  }
-
-  /**
-   * 所持金を更新
-   * @param gold 新しい所持金
-   */
-  updateGold(gold: number): void {
-    this.setGold(gold);
-  }
-
-  /**
-   * コンポーネントを破棄
-   */
-  destroy(): void {
-    if (this.backButton) {
-      this.backButton.destroy();
-      this.backButton = null;
-    }
-    if (this.goldText) {
-      this.goldText.destroy();
-      this.goldText = null;
-    }
-    this.container.destroy();
-  }
-}
+export { ShopHeader } from '@scenes/components/shop/ShopHeader';
