+++ atelier-guild-rank/src/presentation/scenes/components/title/TitleLogo.ts
@@ -1,89 +1,10 @@
 /**
- * TitleLogoコンポーネント
- * TASK-0058 TitleSceneリファクタリング
+ * TitleLogo（後方互換性用再エクスポート）
  *
  * @description
- * タイトルシーンのロゴ部分を表示するコンポーネント
- * - タイトル「ATELIER GUILD」
- * - サブタイトル「錬金術師ギルド」
- * - バージョン情報
+ * 実体は scenes/components/title/TitleLogo.ts に移動済み。
+ * 後方互換性のため再エクスポートを提供する。
+ *
+ * 新規コードでは @scenes/components/title を使用すること。
  */
-
-import { BaseComponent } from '@presentation/ui/components/BaseComponent';
-import { THEME } from '@presentation/ui/theme';
-import { TITLE_LAYOUT, TITLE_STYLES, TITLE_TEXT } from './types';
-
-/**
- * TitleLogoコンポーネント
- * タイトル画面のロゴ・サブタイトル・バージョン情報を担当
- */
-export class TitleLogo extends BaseComponent {
-  /** タイトルテキスト */
-  private titleText: Phaser.GameObjects.Text | null = null;
-
-  /** サブタイトルテキスト */
-  private subtitleText: Phaser.GameObjects.Text | null = null;
-
-  /** バージョンテキスト */
-  private versionText: Phaser.GameObjects.Text | null = null;
-
-  /**
-   * コンポーネントを作成
-   */
-  create(): void {
-    // タイトルテキスト
-    this.titleText = this.scene.add.text(this.container.x, TITLE_LAYOUT.TITLE_Y, TITLE_TEXT.TITLE, {
-      fontFamily: THEME.fonts.primary,
-      fontSize: TITLE_STYLES.TITLE_FONT_SIZE,
-      color: TITLE_STYLES.TITLE_COLOR,
-    });
-    this.titleText.setOrigin(0.5);
-
-    // サブタイトルテキスト
-    this.subtitleText = this.scene.add.text(
-      this.container.x,
-      TITLE_LAYOUT.SUBTITLE_Y,
-      TITLE_TEXT.SUBTITLE,
-      {
-        fontFamily: THEME.fonts.primary,
-        fontSize: TITLE_STYLES.SUBTITLE_FONT_SIZE,
-        color: TITLE_STYLES.SUBTITLE_COLOR,
-      },
-    );
-    this.subtitleText.setOrigin(0.5);
-
-    // バージョン情報テキスト
-    const cameraWidth = this.scene.cameras.main.width;
-    const cameraHeight = this.scene.cameras.main.height;
-    this.versionText = this.scene.add.text(
-      cameraWidth - TITLE_LAYOUT.VERSION_OFFSET,
-      cameraHeight - TITLE_LAYOUT.VERSION_OFFSET,
-      TITLE_TEXT.VERSION,
-      {
-        fontFamily: THEME.fonts.primary,
-        fontSize: TITLE_STYLES.VERSION_FONT_SIZE,
-        color: TITLE_STYLES.VERSION_COLOR,
-      },
-    );
-    this.versionText.setOrigin(1, 1);
-  }
-
-  /**
-   * コンポーネントを破棄
-   */
-  destroy(): void {
-    if (this.titleText) {
-      this.titleText.destroy();
-      this.titleText = null;
-    }
-    if (this.subtitleText) {
-      this.subtitleText.destroy();
-      this.subtitleText = null;
-    }
-    if (this.versionText) {
-      this.versionText.destroy();
-      this.versionText = null;
-    }
-    this.container.destroy();
-  }
-}
+export { TitleLogo } from '@scenes/components/title/TitleLogo';
