+++ atelier-guild-rank/src/scenes/components/rankup/index.ts
@@ -1,0 +1,19 @@
+/**
+ * RankUp コンポーネント バレルエクスポート
+ * TASK-0055 RankUpSceneリファクタリング
+ */
+
+export { RankUpHeader } from './RankUpHeader';
+export { RankUpRequirements } from './RankUpRequirements';
+export { RankUpRewards } from './RankUpRewards';
+export { RankUpTestPanel } from './RankUpTestPanel';
+export type {
+  Artifact,
+  Quality,
+  Rank,
+  RankTest,
+  RankTestTask,
+  RankUpReward,
+  TestPanelCallbacks,
+  TestState,
+} from './types';
