+++ atelier-guild-rank/src/scenes/components/shop/index.ts
@@ -1,0 +1,22 @@
+/**
+ * Shop Components エクスポート
+ * TASK-0056 ShopSceneリファクタリング
+ *
+ * @description
+ * ShopScene関連コンポーネントの一括エクスポート
+ */
+
+// コンポーネントのエクスポート
+export { ShopHeader } from './ShopHeader';
+export { ShopItemCard } from './ShopItemCard';
+export { ShopItemGrid } from './ShopItemGrid';
+// 型定義のエクスポート
+export type {
+  IShopItem,
+  OnBackClickCallback,
+  OnItemSelectCallback,
+  OnPurchaseCallback,
+  ShopHeaderConfig,
+  ShopItemCardConfig,
+  ShopItemGridConfig,
+} from './types';
